pyproject.toml
[tool.poetry]
name = "emojidon"
version = "0.0.0"
description = "Emoji synchronizer for Mastodon"
authors = ["Nik Kantar <nik@nkantar.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
rich = "^13.6.0"
python-dotenv = "^1.0.0"
httpx = "^0.25.1"
python-slugify = "^8.0.1"
selenium = "^4.15.2"
typer = {extras = ["all"], version = "^0.9.0"}

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
pdbpp = "^0.10.3"
mypy = "^1.7.0"
pytest = "^7.4.3"
pytest-watch = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
